News badge on discount item and fixes orders on merchants by category view
